app/pyproject.toml
[tool.poetry]
name = "dbd_bot"
version = "0.1.0"
description = ""
authors = ["pep299"]

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.22.9"
line-bot-sdk = "^2.2.1"
tweepy = "^4.9.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-mock = "^3.7.0"
pytest-env = "^0.6.2"
moto = {extras = ["s3"], version = "^3.1.10"}
pytest-freezegun = "^0.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
